Use vitest instead of jest
