+++ package.json
@@ -18,8 +18,8 @@
     "build:ts": "NODE_ENV=production tsc --emitDeclarationOnly -p ./tsconfig.build.json",
     "build:dts": "api-extractor run --local --verbose && tail -n +3 src/globalExtensions.ts >> dist/index.d.ts",
     "dev": "vite serve -c vite.demo.config.js",
-    "unit": "jest",
-    "unit:watch": "jest --watch",
+    "unit": "vitest",
+    "unit:watch": "vitest  --watch",
     "test": "npm run unit",
     "lint": "eslint ./src/index.ts",
     "preversion": "npm run test && npm run release && git add -A dist"
@@ -60,28 +60,25 @@
   ],
   "devDependencies": {
     "@microsoft/api-extractor": "^7.33.5",
-    "@types/jest": "^29.2.0",
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "@typescript-eslint/parser": "^5.41.0",
     "@vitejs/plugin-vue": "^4.2.3",
     "@vue/compiler-sfc": "^3.3.4",
     "@vue/server-renderer": "^3.2.37",
     "@vue/test-utils": "^2.2.1",
-    "@vue/vue3-jest": "^29.1.1",
     "eslint": "^8.26.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-standard": "^5.0.0",
     "eslint-plugin-vue": "^9.13.0",
-    "jest": "^29.2.2",
-    "jest-environment-jsdom": "^29.2.2",
+    "jsdom": "^22.0.0",
     "mitt": "^3.0.0",
     "sass": "^1.57.1",
-    "ts-jest": "^29.0.3",
     "typescript": "^5.0.4",
     "velocity-animate": "^1.5.2",
     "vite": "^4.3.8",
     "vite-plugin-css-injected-by-js": "^3.1.1",
+    "vitest": "^0.31.1",
     "vue": "^3.3.4"
   }
 }
