+++ src/plugin.ts
@@ -1,9 +1,9 @@
 
 import { App } from 'vue';
-import Notifications from './Notifications.vue';
-import { params } from './params';
-import { NotificationsPluginOptions } from './types';
-import { notify } from './notify';
+import Notifications from '@/components/Notifications.vue';
+import { params } from '@/params';
+import { NotificationsPluginOptions } from '@/types';
+import { notify } from '@/notify';
 
 export function install(app: App, args: NotificationsPluginOptions = {}): void {
   Object.entries(args).forEach((entry) => params.set(...entry));
